WIP: move beacon-store fns to BeaconContext, update imports
